Overworld
First attempt at an overworld. It works, although there is a bug when transition to a new overworld level where the FOV doesn't appear to run (blank screen) and some coordinates appear to be a bit weird when moving back to the town (the stairs are visible close to the edge but they should be ~ 25 tiles in).